Remove the unused codes
