+++ common/functions/src/scalars/arithmetics/mod.rs
@@ -21,7 +21,6 @@ mod arithmetic_mul;
 mod arithmetic_negate;
 mod arithmetic_plus;
 mod binary_arithmetic;
-mod interval;
 mod unary_arithmetic;
 mod utils;
